setting tab _ request component UI fixed
